+++ scripts/winesapos-install.sh
@@ -221,7 +221,7 @@ if [[ "${WINESAPOS_BUILD_CHROOT_ONLY}" == "false" ]]; then
         mount -t vfat "${efi_partition}" "${WINESAPOS_INSTALL_DIR}/boot"
     fi
 
-    for i in tmp var/log var/tmp; do
+    for i in tmp var/log; do
         mkdir -p "${WINESAPOS_INSTALL_DIR}"/${i}
         mount tmpfs -t tmpfs -o nodev,nosuid "${WINESAPOS_INSTALL_DIR}/${i}"
     done
@@ -404,8 +404,7 @@ LABEL=winesapos-boot        	/boot     	ext4      	rw,relatime	0 2
 LABEL=WOS-EFI        	/boot/efi 	vfat      	rw,relatime,fmask=0022,dmask=0022,codepage=437,iocharset=iso8859-1,shortname=mixed,errors=remount-ro	0 2" > "${WINESAPOS_INSTALL_DIR}"/etc/fstab
     # Add temporary mounts separately instead of using 'genfstab -P' to avoid extra file systems.
     echo "tmpfs    /tmp    tmpfs    rw,nosuid,nodev,inode64    0 0
-tmpfs    /var/log    tmpfs    rw,nosuid,nodev,inode64    0 0
-tmpfs    /var/tmp    tmpfs    rw,nosuid,nodev,inode64    0 0" >> "${WINESAPOS_INSTALL_DIR}"/etc/fstab
+tmpfs    /var/log    tmpfs    rw,nosuid,nodev,inode64    0 0" >> "${WINESAPOS_INSTALL_DIR}"/etc/fstab
     elif [[ "${WINESAPOS_BOOTLOADER}" == "systemd-boot" ]]; then
         echo "LABEL=winesapos-root        	/         	btrfs     	rw,noatime,nodiratime,commit=300,compress-force=zstd:1,discard,space_cache=v2,subvolid=$(btrfs subvolume show /winesapos | grep "Subvolume ID"  | awk '{print $3}'),subvol=/	0 0
 LABEL=winesapos-root        	/home     	btrfs     	rw,noatime,nodiratime,commit=300,compress-force=zstd:1,discard,space_cache=v2,subvolid=$(btrfs subvolume show /winesapos/home | grep "Subvolume ID"  | awk '{print $3}'),subvol=/home	0 0
@@ -414,8 +413,7 @@ LABEL=WOS-EFI        	/boot 	vfat      	rw,relatime,fmask=0022,dmask=0022,codepa
     fi
     # Add temporary mounts separately instead of using 'genfstab -P' to avoid extra file systems.
     echo "tmpfs    /tmp    tmpfs    rw,nosuid,nodev,inode64    0 0
-tmpfs    /var/log    tmpfs    rw,nosuid,nodev,inode64    0 0
-tmpfs    /var/tmp    tmpfs    rw,nosuid,nodev,inode64    0 0" >> "${WINESAPOS_INSTALL_DIR}"/etc/fstab
+tmpfs    /var/log    tmpfs    rw,nosuid,nodev,inode64    0 0" >> "${WINESAPOS_INSTALL_DIR}"/etc/fstab
     echo "View final /etc/fstab file:"
     cat "${WINESAPOS_INSTALL_DIR}"/etc/fstab
     echo "Saving partition mounts to /etc/fstab complete."