+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Change Btrfs snapshots to be disabled by default.
 - Add documentation on how to resize a Ventoy image.
 - Allow installing COSMIC, GNOME, Hyprland, and/or Sway during the first-time setup.
+- Remove tmpfs mount on '/var/tmp'.
 
 ## 4.4.0 - The Rolling Update